refactor: reuse quiz button as `ActionButton`
