+++ lib/quiz/view/question_view.dart
@@ -3,7 +3,6 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:provider/provider.dart';
 import 'package:quizapp/quiz/cubit/quiz_cubit.dart';
-import 'package:quizapp/quiz/view/widgets.dart';
 import 'package:quizapp/shared/shared.dart';
 import 'package:quizzes_repository/quizzes_repository.dart';
 import 'package:quizapp/l10n/l10n.dart';
@@ -150,7 +149,7 @@ class QuizAnswerDetails extends StatelessWidget {
                         color: Colors.white54,
                       ),
                     ),
-                    QuizButton(
+                    ActionButton(
                       onPressed: () {
                         context.read<QuizCubit>().validateAnswer();
                       },
